DATA_GENERATION/requirements.txt
# Core data
numpy>=1.26.0
pandas>=2.2.0
pyarrow>=15.0.0

# Cloud storage
boto3>=1.34.0            # AWS S3 + MinIO

# Database
psycopg2-binary>=2.9.9   # PostgreSQL driver

# Utilities
python-dotenv>=1.0.1     # .env file loading
click>=8.1.7             # CLI argument parsing
tenacity>=8.2.3          # retry logic for cloud calls

# Dev / testing
pytest>=8.1.0
